# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: Australian Bank Payment Initiation API
  description: |
    BIAN-compliant payment initiation API for Australian banking systems.
    
    Supports all major Australian payment types:
    - NPP (New Payments Platform) - Real-time instant payments
    - BECS (Bulk Electronic Clearing System) - Batch direct entry
    - BPAY - Bill payment system
    - Direct Debit - Recurring collections
    - Domestic Wire Transfers - High-value payments
    - International Wire Transfers - Cross-border payments
    
    Built following BIAN v12.0.0 Payment Initiation service domain standards.
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@bank.com.au
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3232
    description: Development server
  - url: https://paymentinitiation.codecrushers.club
    description: Production server
paths:
...
  /payment-initiation/npp-payments:
    get:
      summary: Query NPP payments
      description: Search and filter NPP payments
      operationId: queryNPPPayments
      tags:
        - NPP Payments
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/PaymentStatus'
        - name: debitAccount
          in: query
          schema:
            type: string
        - name: creditAccount
          in: query
          schema:
            type: string
        - name: minAmount
          in: query
          schema:
            type: number
        - name: maxAmount
          in: query
          schema:
            type: number
        - name: fromDate
          in: query
          schema:
            type: string
            format: date-time
        - name: toDate
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Payments retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryNPPPaymentsResponse'
...
  /payment-initiation/npp-payments/test-scenarios/{scenarioIndex}/generate:
    post:
      summary: Generate test data
      description: Generate test payment data for a specific scenario
      operationId: generateTestData
      tags:
        - Testing
      parameters:
        - name: scenarioIndex
          in: path
          required: true
          schema:
            type: integer
          description: Test scenario index
      responses:
        '200':
          description: Test data generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateTestDataResponse'
components:
  schemas:
...
    PaymentStatus:
      type: string
      enum:
        - INITIATED
        - PENDING_VALIDATION
        - VALIDATED
        - PENDING_AUTHORIZATION
        - AUTHORIZED
        - REJECTED
        - PROCESSING
        - COMPLETED
        - FAILED
        - CANCELLED
        - RETURNED
...
    QueryNPPPaymentsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            paymentInstructions:
              type: array
              items:
                type: object
            totalCount:
              type: integer
            hasMore:
              type: boolean
        correlationId:
          type: string
        timestamp:
          type: string
          format: date-time
...
    GenerateTestDataResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            scenarioIndex:
              type: integer
            testPaymentData:
              type: object
        timestamp:
          type: string
          format: date-time
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (not implemented for testing)
...