# src/routes/payment-initiation.routes.ts
/**
 * Payment Initiation Routes
 * Express routes for all payment types following BIAN patterns
 */

import { Router } from 'express';
import { NPPPaymentController } from '../controllers/npp-payment.controller';

const router = Router();

// Initialize controllers
const nppController = new NPPPaymentController();

// NPP Payment Routes
const nppRoutes = Router();

// BIAN Standard Operations for NPP Payments
nppRoutes.post('/initiate', nppController.initiatePayment);
nppRoutes.put('/:paymentId/update', nppController.updatePayment);
nppRoutes.post('/:paymentId/request', nppController.requestPayment);
nppRoutes.get('/:paymentId/retrieve', nppController.retrievePayment);
nppRoutes.put('/:paymentId/control', nppController.controlPayment);
nppRoutes.post('/:paymentId/exchange', nppController.exchangePayment);

// Query and utility endpoints
nppRoutes.get('/', nppController.queryPayments);
nppRoutes.get('/test-scenarios', nppController.getTestScenarios);
nppRoutes.post('/test-scenarios/:scenarioIndex/generate', nppController.generateTestData);

// Mount NPP routes
router.use('/npp-payments', nppRoutes);

// BECS Payment Routes (placeholder structure for future implementation)
const becsRoutes = Router();
becsRoutes.post('/initiate', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'BECS payment endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
becsRoutes.put('/:paymentId/update', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'BECS payment endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
becsRoutes.post('/:paymentId/request', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'BECS payment endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
becsRoutes.get('/:paymentId/retrieve', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'BECS payment endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
becsRoutes.put('/:paymentId/control', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'BECS payment endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
becsRoutes.post('/:paymentId/exchange', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'BECS payment endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});

router.use('/becs-payments', becsRoutes);

// BPAY Payment Routes (placeholder structure for future implementation)
const bpayRoutes = Router();
bpayRoutes.post('/initiate', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'BPAY payment endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
bpayRoutes.put('/:paymentId/update', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'BPAY payment endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
bpayRoutes.post('/:paymentId/request', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'BPAY payment endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
bpayRoutes.get('/:paymentId/retrieve', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'BPAY payment endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
bpayRoutes.put('/:paymentId/control', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'BPAY payment endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
bpayRoutes.post('/:paymentId/exchange', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'BPAY payment endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});

router.use('/bpay-payments', bpayRoutes);

// Direct Debit Payment Routes (placeholder structure for future implementation)
const directDebitRoutes = Router();
directDebitRoutes.post('/initiate', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'Direct Debit payment endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
directDebitRoutes.put('/:paymentId/update', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'Direct Debit payment endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
directDebitRoutes.post('/:paymentId/request', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'Direct Debit payment endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
directDebitRoutes.get('/:paymentId/retrieve', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'Direct Debit payment endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
directDebitRoutes.put('/:paymentId/control', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'Direct Debit payment endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
directDebitRoutes.post('/:paymentId/exchange', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'Direct Debit payment endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});

router.use('/direct-debit', directDebitRoutes);

// Domestic Wire Transfer Routes (placeholder structure for future implementation)
const domesticWireRoutes = Router();
domesticWireRoutes.post('/initiate', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'Domestic Wire transfer endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
domesticWireRoutes.put('/:paymentId/update', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'Domestic Wire transfer endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
domesticWireRoutes.post('/:paymentId/request', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'Domestic Wire transfer endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
domesticWireRoutes.get('/:paymentId/retrieve', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'Domestic Wire transfer endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
domesticWireRoutes.put('/:paymentId/control', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'Domestic Wire transfer endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
domesticWireRoutes.post('/:paymentId/exchange', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'Domestic Wire transfer endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});

router.use('/domestic-wires', domesticWireRoutes);

// International Wire Transfer Routes (placeholder structure for future implementation)
const internationalWireRoutes = Router();
internationalWireRoutes.post('/initiate', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'International Wire transfer endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
internationalWireRoutes.put('/:paymentId/update', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'International Wire transfer endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
internationalWireRoutes.post('/:paymentId/request', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'International Wire transfer endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
internationalWireRoutes.get('/:paymentId/retrieve', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'International Wire transfer endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
internationalWireRoutes.put('/:paymentId/control', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'International Wire transfer endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});
internationalWireRoutes.post('/:paymentId/exchange', (req, res) => {
  res.status(501).json({
    success: false,
    errors: [{
      errorCode: 'NOT_IMPLEMENTED',
      errorDescription: 'International Wire transfer endpoints are not yet implemented'
    }],
    timestamp: new Date().toISOString()
  });
});

router.use('/international-wires', internationalWireRoutes);

// Health check and service information endpoints
router.get('/health', (req, res) => {
  res.status(200).json({
    success: true,
    data: {
      status: 'healthy',
      serviceDomain: 'Payment Initiation',
      version: '1.0.0',
      bianCompliant: true,
      supportedPaymentTypes: [
        'NPP_INSTANT',
        'NPP_PAYID',
        'BECS_DIRECT_ENTRY',
        'BPAY_PAYMENT',
        'DIRECT_DEBIT',
        'DOMESTIC_WIRE',
        'INTERNATIONAL_WIRE'
      ],
      implementedPaymentTypes: [
        'NPP_INSTANT',
        'NPP_PAYID'
      ]
    },
    timestamp: new Date().toISOString()
  });
});

router.get('/service-info', (req, res) => {
  res.status(200).json({
    success: true,
    data: {
      serviceInstanceId: 'AUS-BANK-PI-001',
      serviceDomainVersion: '12.0.0',
      bianStandardVersion: 'v12.0.0',
      complianceLevel: 'FULL',
      supportedOperations: [
        'INITIATE',
        'UPDATE',
        'REQUEST',
        'RETRIEVE',
        'CONTROL',
        'EXCHANGE'
      ],
      australianPaymentSystems: {
        npp: 'New Payments Platform - Real-time payments',
        becs: 'Bulk Electronic Clearing System - Batch payments',
        bpay: 'BPAY - Bill payment system',
        directDebit: 'Direct Debit - Recurring payments',
        domesticWire: 'Domestic Wire Transfers - High value payments',
        internationalWire: 'International Wire Transfers - Cross-border payments'
      }
    },
    timestamp: new Date().toISOString()
  });
});

export default router;